requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.12.5
PyJWT==2.9.0
PyYAML==6.0.3
python-multipart==0.0.9

